requirements.txt
albumentations==0.4.3
matplotlib==3.3.4
numpy>=1.11.1
opencv_python==4.8.0.76
openslide_python==1.1.2
pandas==1.4.4
scikit_image==0.18.1
scikit_learn==1.1.2
scipy==1.7.0
scikit-image==0.18.1
tables==3.6.1
tifffile==2021.11.2
torch==2.0.0
torchsummary==1.5.1
torchvision==0.15.1
tensorboard==2.14.0
tqdm==4.62.3
ttach==0.0.3
flake8==4.0.1
filelock==3.8.0
platformdirs==2.5.2
pytest>=6.2.5
